+++ src/client.rs
@@ -35,7 +35,7 @@ pub static PREPARED_STATEMENT_COUNTER: Lazy<Arc<AtomicUsize>> =
     Lazy::new(|| Arc::new(AtomicUsize::new(0)));
 pub static CLIENT_COUNTER: Lazy<Arc<AtomicUsize>> = Lazy::new(|| Arc::new(AtomicUsize::new(0)));
 // Ignore deallocate queries from pgx.
-static QUERY_DEALLOCATE: &[u8] = "deallocate \"".as_bytes();
+static QUERY_DEALLOCATE: &[u8] = "deallocate ".as_bytes();
 
 /// Type of connection received from client.
 enum ClientConnectionType {